core: Move mapped storage primitives
Mapped byte and record storage backs LineBuffer and batch matching without
materializing large files on the Python heap. The primitive belongs with
the neutral buffer and matching model rather than Git or filesystem
helpers.

Keeping the module in utils made core.buffer import a utility module while
repository buffer utilities import LineBuffer. That gave the storage
primitive a package home that obscured the core/utils boundary.

This commit addresses that by moving mapped_storage into core and updating
core and batch callers to import it from there. The old utils module is
removed instead of left as a compatibility shim.

Follow-up test commits will move the tests and add a boundary guard for the
mapped storage package location.
