+++ src/git_stage_batch/core/mapped_storage.py
